requirements.txt
appnope==0.1.0
decorator==4.0.6
dj-database-url==0.4.0
Django==1.9.2
django-extensions==1.6.1
djangorestframework==3.3.2
funcsigs==0.4
gnureadline==6.3.3
gunicorn==19.4.5
ipdb==0.8.1
ipython==4.1.1
ipython-genutils==0.1.0
mock==1.3.0
path.py==8.1.2
pbr==1.8.1
pexpect==4.0.1
pickleshare==0.6
psycopg2==2.6.1
ptyprocess==0.5.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.1.0
whitenoise==2.0.6
